requirements.txt
streamlit
pillow
pytesseract
PyPDF2
google-generativeai
